requirements.txt
numpy>=1.21.0
PyOpenGL>=3.1.0
pandas>=1.3.0
pyyaml>=5.4.1
matplotlib>=3.4.0
# Dependencias para visualización 3D
pyqt5>=5.15.0
# Dependencias para análisis de datos
scipy>=1.7.0
# Dependencias para logging
logging>=0.5.1.2 